feat: add validators on all controllers

+++ src/controllers/ticketController.js
@@ -1,14 +1,31 @@
 const ticketRepository = require('../repositories/ticketRepository');
 const ValidationContract = require('../util/validators');
+const attendantController = require('./attendantController');
 
 exports.createTicket = async (req, res) => {
   let validators = new ValidationContract();
   validators.isRequired(req.body.idAttendant, 'idAttendant is required');
   validators.isRequired(req.body.title, 'title is required');
   validators.isRequired(req.body.idClient, 'idClient is required');
+  validators.isObjectIdValid(
+    req.body.idAttendant,
+    `idAttendant: "${req.body.idAttendant}" is not a ObjectId valid.`,
+  );
+  validators.isObjectIdValid(
+    req.body.idClient,
+    `idClient: "${req.body.idClient}" is not a ObjectId valid.`,
+  );
 
   try {
     if (validators.isValid()) {
+      const responseAttendant = await attendantController.findAttendantById(
+        req.body.idAttendant,
+      );
+      if (responseAttendant.status !== 200) {
+        res.status(responseAttendant.status).send(responseAttendant.text);
+        return;
+      }
+
       await ticketRepository.createTicket(req.body);
       res.status(201).send('Ticket created!');
     } else {
@@ -27,7 +44,7 @@ exports.findAllTickets = async (req, res) => {
   try {
     const tickets = await ticketRepository.findAllTickets();
     if (tickets == null || tickets.length == 0) {
-      res.status(204).send('No tickets found');
+      res.status(204).send();
     } else {
       res.status(200).send(tickets);
     }
@@ -39,27 +56,82 @@ exports.findAllTickets = async (req, res) => {
 };
 
 exports.findTicketById = async (req, res) => {
+  let validators = new ValidationContract();
   const ticketId = req.params.id;
-  if (ticketId == null) {
-    res.status(400).send('ticketId is required');
+  validators.isRequired(ticketId, 'ticketId is required');
+  validators.isObjectIdValid(
+    ticketId,
+    `ticketId: "${ticketId}" is not a ObjectId valid.`,
+  );
+  try {
+    if (validators.isValid()) {
+      const ticket = await ticketRepository.findTicketById(ticketId);
+      if (!ticket) {
+        res.status(204).send();
+        return;
+      }
+      res.status(200).send(ticket);
+    } else {
+      res.status(400).send({
+        errors: validators.getErrors(),
+      });
+    }
+  } catch (error) {
+    console.log(error);
+    res.status(500).send({
+      message: 'Server error.',
+    });
   }
-  const ticket = ticketRepository.findTicketById(ticketId);
-
-  if (!ticket) {
-    res.status(404).send();
-  }
-
-  res.status(200).send(ticket);
 };
 
 exports.updateTicketById = async (req, res) => {
+  let validators = new ValidationContract();
   const ticketId = req.params.id;
-  await ticketRepository.updateTicketById(ticketId, req.body);
-  res.status(200).send('Ticket updated', req.body);
+  validators.isRequired(ticketId, 'ticketId is required');
+  validators.isObjectIdValid(
+    ticketId,
+    `ticketId: "${ticketId}" is not a ObjectId valid.`,
+  );
+
+  try {
+    if (validators.isValid()) {
+      await ticketRepository.updateTicketById(ticketId, req.body);
+      res.status(200).send('Ticket updated');
+    } else {
+      res.status(400).send({
+        errors: validators.getErrors(),
+      });
+    }
+  } catch (error) {
+    console.log(error);
+    res.status(500).send({
+      message: 'Server error.',
+    });
+  }
 };
 
 exports.deleteTicketById = async (req, res) => {
+  let validators = new ValidationContract();
   const ticketId = req.params.id;
-  await ticketRepository.deleteTicketById(ticketId);
-  res.status(204).send('Ticket deleted', req.body);
+  validators.isRequired(ticketId, 'ticketId is required');
+  validators.isObjectIdValid(
+    ticketId,
+    `ticketId: "${ticketId}" is not a ObjectId valid.`,
+  );
+
+  try {
+    if (validators.isValid()) {
+      await ticketRepository.deleteTicketById(ticketId);
+      res.status(204).send();
+    } else {
+      res.status(400).send({
+        errors: validators.getErrors(),
+      });
+    }
+  } catch (error) {
+    console.log(error);
+    res.status(500).send({
+      message: 'Server error.',
+    });
+  }
 };
